Restore develop

+++ src/main/java/com/intellecteu/onesource/integration/dto/SettlementDto.java
@@ -1,8 +1,8 @@
 package com.intellecteu.onesource.integration.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.intellecteu.onesource.integration.model.PartyRole;
-import com.intellecteu.onesource.integration.model.SettlementStatus;
+import com.intellecteu.onesource.integration.model.onesource.PartyRole;
+import com.intellecteu.onesource.integration.model.onesource.SettlementStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -17,7 +17,7 @@ public class SettlementDto {
     @JsonProperty("partyRole")
     private PartyRole partyRole;
     @JsonProperty("instructionId")
-    private Integer instructionId;
+    private Long instructionId;
     @JsonProperty("settlementStatus")
     private SettlementStatus settlementStatus;
     @JsonProperty("instruction")
